refactor:etc폴더명변경&hiddenUser추가
