+++ src/main/java/com/osunji/melog/calendar/Calendar.java
@@ -1,4 +1,4 @@
-package com.osunji.melog.etc;
+package com.osunji.melog.calendar;
 
 import java.time.LocalDate;
